Remove intakeCommon
