+++ client/app/intake/components/IssueCounter.jsx
