+++ client/app/intake/pages/addIssues.jsx
@@ -1,12 +1,13 @@
 import _ from 'lodash';
-import React from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
+import React from 'react';
 
 import AddIssuesModal from '../components/AddIssuesModal';
 import NonRatedIssueModal from '../components/NonRatedIssueModal';
 import Button from '../../components/Button';
-import { FORM_TYPES } from '../../intakeCommon/constants';
+import RequestIssuesUpdateErrorAlert from '../../intakeEdit/components/RequestIssuesUpdateErrorAlert';
+import { FORM_TYPES } from '../constants';
 import { formatDate } from '../../util/DateUtil';
 import { formatAddedIssues, getAddIssuesFields } from '../util/issues';
 
@@ -18,12 +19,13 @@ class AddIssuesPage extends React.PureComponent {
     const {
       intakeForms,
       formType,
-      veteran
+      veteran,
+      responseErrorCode,
     } = this.props;
 
     const selectedForm = _.find(FORM_TYPES, { key: formType });
-    const veteranInfo = `${veteran.name} (${veteran.fileNumber})`;
     const intakeData = intakeForms[selectedForm.key];
+    const veteranInfo = `${veteran.name} (${veteran.fileNumber})`;
 
     const issuesComponent = () => {
       let issues = formatAddedIssues(intakeData);
@@ -92,6 +94,10 @@ class AddIssuesPage extends React.PureComponent {
       }
       <h1 className="cf-txt-c">Add Issues</h1>
 
+      { responseErrorCode &&
+        <RequestIssuesUpdateErrorAlert responseErrorCode={responseErrorCode} />
+      }
+
       <Table
         columns={columns}
         rowObjects={rowObjects}
@@ -108,7 +114,8 @@ export const IntakeAddIssuesPage = connect(
       appeal
     },
     formType: intake.formType,
-    veteran: intake.veteran
+    veteran: intake.veteran,
+    responseErrorCode: null
   }),
   (dispatch) => bindActionCreators({
     toggleAddIssuesModal,
@@ -125,7 +132,9 @@ export const EditAddIssuesPage = connect(
       appeal: state
     },
     formType: state.formType,
-    veteran: state.veteran
+    veteran: state.veteran,
+    requestStatus: state.requestStatus.requestIssuesUpdate,
+    responseErrorCode: state.responseErrorCode
   }),
   (dispatch) => bindActionCreators({
     toggleAddIssuesModal,
