+++ client/app/intake/pages/rampElection/finish.jsx
@@ -6,8 +6,7 @@ import Checkbox from '../../../components/Checkbox';
 import Alert from '../../../components/Alert';
 import Table from '../../../components/Table';
 import { Redirect } from 'react-router-dom';
-import { PAGE_PATHS, INTAKE_STATES } from '../../constants';
-import { REQUEST_STATE } from '../../../intakeCommon/constants';
+import { PAGE_PATHS, INTAKE_STATES, REQUEST_STATE } from '../../constants';
 import { connect } from 'react-redux';
 import { completeIntake, confirmFinishIntake } from '../../actions/rampElection';
 import { bindActionCreators } from 'redux';
