+++ client/app/intake/pages/rampElection/review.jsx
@@ -9,8 +9,7 @@ import Button from '../../../components/Button';
 import _ from 'lodash';
 import { submitReview } from '../../actions/rampElection';
 import { setReceiptDate, setOptionSelected } from '../../actions/intake';
-import { PAGE_PATHS, INTAKE_STATES, REVIEW_OPTIONS } from '../../constants';
-import { REQUEST_STATE } from '../../../intakeCommon/constants';
+import { PAGE_PATHS, INTAKE_STATES, REVIEW_OPTIONS, REQUEST_STATE } from '../../constants';
 import { getIntakeStatus } from '../../selectors';
 import ErrorAlert from '../../components/ErrorAlert';
