+++ client/app/intake/pages/rampRefiling/review.jsx
@@ -8,8 +8,7 @@ import Button from '../../../components/Button';
 import Alert from '../../../components/Alert';
 import { Redirect } from 'react-router-dom';
 import _ from 'lodash';
-import { PAGE_PATHS, INTAKE_STATES, REVIEW_OPTIONS } from '../../constants';
-import { REQUEST_STATE } from '../../../intakeCommon/constants';
+import { PAGE_PATHS, INTAKE_STATES, REVIEW_OPTIONS, REQUEST_STATE } from '../../constants';
 import { setAppealDocket, submitReview, confirmIneligibleForm } from '../../actions/rampRefiling';
 import { setReceiptDate, setOptionSelected } from '../../actions/intake';
 import { toggleIneligibleError } from '../../util';
