+++ client/app/intake/pages/selectForm.jsx
@@ -6,8 +6,7 @@ import { Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { setFormType, clearSearchErrors } from '../actions/intake';
-import { PAGE_PATHS } from '../constants';
-import { FORM_TYPES } from '../../intakeCommon/constants';
+import { PAGE_PATHS, FORM_TYPES } from '../constants';
 import _ from 'lodash';
 
 class SelectForm extends React.PureComponent {
