+++ client/app/intake/reducers/intake.js
@@ -1,5 +1,4 @@
-import { ACTIONS } from '../constants';
-import { REQUEST_STATE } from '../../intakeCommon/constants';
+import { ACTIONS, REQUEST_STATE } from '../constants';
 import { update } from '../../util/ReducerUtil';
 import { formatDateStr } from '../../util/DateUtil';
 import _ from 'lodash';
