+++ client/app/intakeEdit/constants.js
@@ -1,6 +1,5 @@
 export const PAGE_PATHS = {
   BEGIN: '/',
-  SELECT_ISSUES: '/select_issues',
   CANCEL_ISSUES: '/cancel'
 };
 
@@ -9,7 +8,7 @@ export const ENDPOINT_NAMES = {
 };
 
 export const ACTIONS = {
-  SET_ISSUE_SELECTED: 'SET_ISSUE_SELECTED',
+  NO_ISSUES_SELECTED_ERROR: 'NO_ISSUES_SELECTED_ERROR',
   REQUEST_ISSUES_UPDATE_START: 'REQUEST_ISSUES_UPDATE_START',
   REQUEST_ISSUES_UPDATE_SUCCEED: 'REQUEST_ISSUES_UPDATE_SUCCEED',
   REQUEST_ISSUES_UPDATE_FAIL: 'REQUEST_ISSUES_UPDATE_FAIL'
